TVB-2531 Changes for compatibility with PostgreSQL

+++ framework_tvb/tvb/adapters/datatypes/db/fcd.py
@@ -52,8 +52,8 @@ class FcdIndex(DataTypeMatrix):
         # type: (Fcd)  -> None
         super(FcdIndex, self).fill_from_has_traits(datatype)
         self.array_data_min, self.array_data_max, self.array_data_mean = from_ndarray(datatype.array_data)
-        self.array_data_min = numpy.float64(self.array_data_min)
-        self.array_data_mean = numpy.float64(self.array_data_mean)
-        self.array_data_max = numpy.float64(self.array_data_max)
+        self.array_data_min = self.array_data_min
+        self.array_data_mean = self.array_data_mean
+        self.array_data_max = self.array_data_max
         self.labels_ordering = json.dumps(datatype.labels_ordering)
         self.source_gid = datatype.source.gid.hex
